fixing db connection

+++ backend/app/db/db_connection.py
@@ -1,16 +1,10 @@
-from pymongo import MongoClient
-from pymongo.server_api import ServerApi
-from os import environ
+from . import connect
 
-password = environ.get("DB_PASSWORD")
-username = environ.get("DB_USERNAME")
-appName = environ.get("APP_NAME")
 
-_uri = f"mongodb+srv://{username}:{password}@{appName}.lbpiqmb.mongodb.net/?appName={appName}"
-_client = MongoClient(_uri, server_api=ServerApi('1'))
-_db = _client["my_db"]
-
-collection = _db["my_collection"]
+collection, coll_exists = connect()
 
 def reset_collection():
-    collection.delete_many({})
+    if coll_exists:
+        collection.delete_many({})
+        print("Collection reset")
+    return
